+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [`tikv`] Fixed `store.WithEmptyValue` support when using compression that was not compressing the "formatted" value that should have been sent to TiKV.
 
 ### Changed
+- [`tivk`] Upgraded to `tikv-client/v2` version, here the changes related to that:
+ - The raw max scan limit dsn query parameter `tikv_raw_max_scan_limit=<value>` now applies globally to all instances. This means if in the use application, multiple DSN for TiKV are provided, the last one with `tikv_raw_max_scan_limit` wins.
+ - The raw max batch put size dsn query parameter `tikv_raw_max_batch_put_size=<value>` has been removed (the client keeps it private now).
+ - The raw batch pair count dsn query parameter `tikv_raw_batch_pair_count=<value>` has been removed (the client keeps it private now).
 - [`core`] Now supporting key-only iteration for `BatchPrefix`, `Prefix` and `Scan` calls.
 - [`core`] **BREAKING** Added `options ...store.ReadOption` options to `store.KVStore#BatchPrefix`.
 - [`core`] **BREAKING** Added `options ...store.ReadOption` options to `store.KVStore#Prefix`.